Fix Android permission and Android manifest for Microbit board
The Android manifest file was registering the wrong vendor/product id (it was in hex while Android expects decimal values).
Now the app won't need to ask permission while running.

Also, the example application is now receiving a callback when a compatible board is connected and the list of devices is refreshed.