Fix initialization of signed long variables
The values inserted on initialization of long variables ar treated as integer.
So, it will automaticly makes a convertion from signed integer to signed long if the value is inserted without the letter "L" at the end of the value.

An  "long var = 0x80000000" would be converted to "0xFFFFFFFF80000000" (long) leading to errors when bitwise operations is performed.