Fix infinite loop if board in an unexpected state
When the board is in an unexpected state we might not be able to flash with jOCD.

This is a known issue but can be easily fixed using openOCD.

This fix avoid locking the app in an infinite loop while waiting for an specific state.
It includes a counter and throws and exception if the board do not answer as expected.